Some questline fixups
Added grunt 103 to the list of eligible grunts for the quest.
Changed the QuestLineStepCompletedRequirement to quest 0. rainbowInvasion is the first quest in the chain, this makes it quest 0, GameConstants.AchievementOption.less means this quest is NOT cleared, so the requirement is "the first quest in the chain is not cleared", whereas before it was "the second quest in the chain is not cleared".