Rainbow Rocket for Aether Foundation
Fully Rainbow Rocketify the Aether Foundation.

A new fancy overlay becomes active during the Defeat Rainbow Rocket questline. Intentionally placed above all of the other Aether Foundation SVG so that that stuff is still laid on top of this overlay.
Also moved all the Rainbow Rocket temporarybattles to the Aether Foundation building.

Also features a fix to QuestLineHelper.ts to make the game actually load, a fix to the Rainbow Leaders requirements so that they activate at the right time, and a typofix on one grunt's returntown.